backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
deepgram-sdk==3.0.0
groq==0.4.0
edge-tts==6.1.9
python-dotenv==1.0.0
